REFACTOR: ScoreCalculator 생성으로 Service 계층 책임 분리
- 점수 계산을 담당하는 util 서비스 ScoreCalculator 생성
- Service 계층의 메서드 내 점수 계산 로직을 ScoreCalculator로 위임
